Only test/build Linux-specific parts of wsutil and ptyutil on Linux (#74)
This avoids running the tests for the Linux-specific parts of this on
non-Linux OSes, so that `github.com/canonical/pebble` is (for example)
importable in Juju and `go test ./...` runs on Windows and macOS.

The `_linux.go` suffix is also how the LXD codebase does it.
